Add new textures, create random namespace

+++ src/Asteroid.cpp
@@ -4,8 +4,8 @@ Asteroid::Asteroid() : velocity(0.0f), angularVelocity(0.0f)
 {
 }
 
-void Asteroid::update(double deltaTime)
+void Asteroid::update(float deltaTime)
 {
-    rotation += angularVelocity * (float)deltaTime;
-    position += velocity * (float)deltaTime;
+    rotation += angularVelocity * deltaTime;
+    position += velocity * deltaTime;
 }
